DTFS2-4174 - split up TFM tasks functions
