+++ trade-finance-manager-api/src/v1/controllers/tasks.controller.js
@@ -1,37 +1,29 @@
-const { getTime } = require('date-fns');
 const api = require('../api');
 const CONSTANTS = require('../../constants');
-const getAssigneeFullName = require('./get-assignee-full-name');
+const getAssigneeFullName = require('../helpers/get-assignee-full-name');
 const {
-  previousTaskIsComplete,
-  taskCanBeEditedWithoutPreviousTaskComplete,
   isFirstTaskInFirstGroup,
   getGroupById,
   getGroupByTitle,
   getTaskInGroup,
   getTaskInGroupByTitle,
-  handleTaskEditFlagAndStatus,
+  taskIsCompletedImmediately,
+  isAdverseHistoryTaskIsComplete,
+  shouldUpdateDealStage,
 } = require('../helpers/tasks');
+const mapTaskObject = require('../tasks/map-task-object');
+const mapTaskHistoryObject = require('../tasks/map-task-history-object');
+const {
+  previousTaskIsComplete,
+  taskCanBeEditedWithoutPreviousTaskComplete,
+  handleTaskEditFlagAndStatus,
+} = require('../tasks/tasks-edit-logic');
 const sendUpdatedTaskEmail = require('./task-emails');
 const mapSubmittedDeal = require('../mappings/map-submitted-deal');
 
-const updateHistory = ({
-  taskId,
-  groupId,
-  statusFrom,
-  statusTo,
-  assignedUserId,
-  updatedBy,
-}) => ({
-  taskId,
-  groupId,
-  statusFrom,
-  statusTo,
-  assignedUserId,
-  updatedBy,
-  timestamp: getTime(new Date()),
-});
-
+/**
+* Update a task in the task's group.
+* */
 const updateTask = (allTaskGroups,taskUpdate) =>
   allTaskGroups.map((tGroup) => {
     let group = tGroup;
@@ -57,48 +49,9 @@ const updateTask = (allTaskGroups,taskUpdate) =>
   });
 
 /**
-* Generate timestamps:
-* - lastEdited
-* - dateStarted
-* - dateCompleted
-* */
-const generateTaskDates = (statusFrom, statusTo) => {
-  const dates = {
-    lastEdited: getTime(new Date()),
-  };
-
-  if (statusFrom === CONSTANTS.TASKS.STATUS.TO_DO) {
-    dates.dateStarted = getTime(new Date());
-  }
-
-  if (statusTo === CONSTANTS.TASKS.STATUS.COMPLETED) {
-    dates.dateCompleted = getTime(new Date());
-  }
-
-  return dates;
-};
-
-/**
- * Get the Adverse History group and
- * check if the Adverse History Check task is completed.
+ * Update all tasks canEdit flag and status
+ * Depending on what task has been changed.
  * */
-const isAdverseHistoryTaskIsComplete = (allTaskGroups) => {
-  const adverseGroup = getGroupByTitle(allTaskGroups, CONSTANTS.TASKS.GROUP_TITLES.ADVERSE_HISTORY);
-
-  if (adverseGroup) {
-    const adverseTaskTitle = CONSTANTS.TASKS.MIA_ADVERSE_HISTORY_GROUP_TASKS.COMPLETE_ADVERSE_HISTORY_CHECK;
-
-    const adverseTask = getTaskInGroupByTitle(adverseGroup.groupTasks, adverseTaskTitle);
-    
-    if (adverseTask
-      && adverseTask.status === CONSTANTS.TASKS.STATUS.COMPLETED) {
-      return true;
-    }
-  }
-
-  return false;
-};
-
 const updateTasksCanEdit = async (allTaskGroups, groupId, taskUpdate, deal, urlOrigin) => {
   let taskGroups = allTaskGroups.map((tGroup) => {
     let group = tGroup;
@@ -133,7 +86,6 @@ const updateTasksCanEdit = async (allTaskGroups, groupId, taskUpdate, deal, urlO
    * If 'Adverse History Check' task is completed
    * All tasks in the Underwriting Group Task become unlocked/able to be started
    * */
-
   const canUnlockUnderWritingGroupTasks = isAdverseHistoryTaskIsComplete(taskGroups);
 
   if (canUnlockUnderWritingGroupTasks) {
@@ -181,67 +133,17 @@ const updateTasksCanEdit = async (allTaskGroups, groupId, taskUpdate, deal, urlO
   return taskGroups;
 };
 
-const isMIAdeal = (submissionType) =>
-  submissionType === CONSTANTS.DEALS.SUBMISSION_TYPE.MIA;
-
-const taskIsCompletedImmediately = (statusFrom, statusTo) => {
-  if (statusFrom === CONSTANTS.TASKS.STATUS.TO_DO
-    && statusTo === CONSTANTS.TASKS.STATUS.COMPLETED) {
-    return true;
-  }
-
-  return false;
-};
-
-const shouldUpdateDealStage = (submissionType, taskId, groupId, statusFrom, statusTo) => {
-  const miaDeal = isMIAdeal(submissionType);
-  const firstTaskInFirstGroup = isFirstTaskInFirstGroup(taskId, groupId);
-  const taskCompletedImmediately = taskIsCompletedImmediately(statusFrom, statusTo);
-
-  if (miaDeal
-    && firstTaskInFirstGroup
-    && (statusTo === CONSTANTS.TASKS.STATUS.IN_PROGRESS || taskCompletedImmediately)) {
-    return true;
-  }
-
-  return false;
-};
-
 /**
- * Construct a new object from inputted/requested data
+ * Function that is first triggered. This:
+ * - Gets the deal and all tasks.
+ * - Maps the taskUpdate input into schema format, adding dates.
+ * - Finds the group the task belongs to and updates it.
+ * - Maps over all tasks and updates their status/canEdit flag.
+ * - Updates tfm.history.tasks.
+ * - Change the TFM dealStage (if deal is MIA and taskUpdate is first task)
+ * - Updates the deal.
  * */
-const mapTaskObject = async (originalTask, updateInput) => {
-  const statusFrom = originalTask.status;
-
-  const {
-    id: taskIdToUpdate,
-    groupId,
-    assignedTo,
-    status: statusTo,
-    updatedBy,
-    urlOrigin,
-  } = updateInput;
-
-  const { userId: assignedUserId } = assignedTo;
-
-  const newAssigneeFullName = await getAssigneeFullName(assignedUserId);
-
-  const taskObj = {
-    ...originalTask,
-    id: taskIdToUpdate,
-    groupId,
-    status: statusTo,
-    previousStatus: statusFrom,
-    assignedTo: {
-      userFullName: newAssigneeFullName,
-      userId: assignedUserId,
-    },
-    ...generateTaskDates(statusFrom, statusTo),
-  };
-
-  return taskObj;
-};
-
+// TODO ^^^ ADD send emails.
 const updateTfmTask = async (dealId, taskUpdate) => {
   const unmappedDeal = await api.findOneDeal(dealId);
   const deal = mapSubmittedDeal(unmappedDeal);
@@ -277,6 +179,8 @@ const updateTfmTask = async (dealId, taskUpdate) => {
      * Map over every task in all groups and
      * update other tasks so they can be started.
      * E.g If task 1 is completed, task 2 can then be started.
+     * Some other special conditions are in here:
+     * - e.g if X task is completed, an entire group of tasks can be started.
      * */
     const modifiedTasksWithEditStatus = await updateTasksCanEdit(
       modifiedTasks,
@@ -291,7 +195,7 @@ const updateTfmTask = async (dealId, taskUpdate) => {
      * */
     const tfmHistoryUpdate = {
       tasks: [
-        updateHistory({
+        mapTaskHistoryObject({
           taskId: taskIdToUpdate,
           groupId,
           statusFrom,
@@ -313,7 +217,9 @@ const updateTfmTask = async (dealId, taskUpdate) => {
     };
 
     /**
-     * Check if we need to update the deal stage
+     * If the deal is MIA
+     * and the first task status is being changed to in progress or is completed immediately,
+     * the TFM deal stage needs to change to 'in progress'.
      * */
     const updateDealStage = shouldUpdateDealStage(
       deal.submissionType,
@@ -338,111 +244,9 @@ const updateTfmTask = async (dealId, taskUpdate) => {
   return originalTask;
 };
 
-/**
- * assignTeamTasksToOneUser is no longer consumed.
- * Leaving this here for potential future use.
- * */
-const assignTeamTasksToOneUser = async (dealId, teamIds, userId) => {
-  const deal = await api.findOneDeal(dealId);
-  const allTaskGroups = deal.tfm.tasks;
-
-  const newAssigneeFullName = await getAssigneeFullName(userId);
-
-  let modifiedTasks = allTaskGroups;
-
-  modifiedTasks = modifiedTasks.map((tGroup) => {
-    let group = tGroup;
-
-    group = {
-      ...group,
-      groupTasks: group.groupTasks.map((t) => {
-        let task = t;
-
-        if (teamIds.includes(task.team.id)) {
-          task = {
-            ...task,
-            assignedTo: {
-              userFullName: newAssigneeFullName,
-              userId,
-            },
-          };
-        }
-
-        return task;
-      }),
-    };
-
-    return group;
-  });
-
-  const tfmDealUpdate = {
-    tfm: {
-      tasks: modifiedTasks,
-    },
-  };
-
-  await api.updateDeal(dealId, tfmDealUpdate);
-
-  return modifiedTasks;
-};
-
-const assignGroupTasksToOneUser = async (dealId, groupTitlesToAssign, userId) => {
-  const deal = await api.findOneDeal(dealId);
-  const allTaskGroups = deal.tfm.tasks;
-
-  const newAssigneeFullName = await getAssigneeFullName(userId);
-
-  let modifiedTasks = allTaskGroups;
-
-  modifiedTasks = modifiedTasks.map((tGroup) => {
-    let group = tGroup;
-
-    group = {
-      ...group,
-      groupTasks: group.groupTasks.map((t) => {
-        let task = t;
-
-        /**
-         * Use group title instead of id.
-         * This ensures that the functionality will still work
-         * if the ordering is changed.
-         * */
-        if (groupTitlesToAssign.includes(group.groupTitle)) {
-          task = {
-            ...task,
-            assignedTo: {
-              userFullName: newAssigneeFullName,
-              userId,
-            },
-          };
-        }
-
-        return task;
-      }),
-    };
-
-    return group;
-  });
-
-  const tfmDealUpdate = {
-    tfm: {
-      tasks: modifiedTasks,
-    },
-  };
-
-  await api.updateDeal(dealId, tfmDealUpdate);
-
-  return modifiedTasks;
-};
-
 module.exports = {
   updateTask,
-  generateTaskDates,
   updateTasksCanEdit,
-  isMIAdeal,
-  taskIsCompletedImmediately,
   shouldUpdateDealStage,
   updateTfmTask,
-  assignTeamTasksToOneUser,
-  assignGroupTasksToOneUser,
 };
