+++ trade-finance-manager-api/src/v1/helpers/get-assignee-full-name.js
