+++ trade-finance-manager-api/src/v1/tasks/assign-group-tasks-to-one-user.js
@@ -1,0 +1,53 @@
+const api = require('../api');
+const getAssigneeFullName = require('../helpers/get-assignee-full-name');
+
+const assignGroupTasksToOneUser = async (dealId, groupTitlesToAssign, userId) => {
+  const deal = await api.findOneDeal(dealId);
+  const allTaskGroups = deal.tfm.tasks;
+
+  const newAssigneeFullName = await getAssigneeFullName(userId);
+
+  let modifiedTasks = allTaskGroups;
+
+  modifiedTasks = modifiedTasks.map((tGroup) => {
+    let group = tGroup;
+
+    group = {
+      ...group,
+      groupTasks: group.groupTasks.map((t) => {
+        let task = t;
+
+        /**
+         * Use group title instead of id.
+         * This ensures that the functionality will still work
+         * if the ordering is changed.
+         * */
+        if (groupTitlesToAssign.includes(group.groupTitle)) {
+          task = {
+            ...task,
+            assignedTo: {
+              userFullName: newAssigneeFullName,
+              userId,
+            },
+          };
+        }
+
+        return task;
+      }),
+    };
+
+    return group;
+  });
+
+  const tfmDealUpdate = {
+    tfm: {
+      tasks: modifiedTasks,
+    },
+  };
+
+  await api.updateDeal(dealId, tfmDealUpdate);
+
+  return modifiedTasks;
+};
+
+module.exports = assignGroupTasksToOneUser;
