+++ trade-finance-manager-api/src/v1/tasks/assign-team-tasks-to-one-user.js
@@ -1,0 +1,54 @@
+const api = require('../api');
+const getAssigneeFullName = require('../helpers/get-assignee-full-name');
+
+/**
+ * assignTeamTasksToOneUser is no longer consumed.
+ * Leaving this here for potential future use.
+ * */
+
+const assignTeamTasksToOneUser = async (dealId, teamIds, userId) => {
+  const deal = await api.findOneDeal(dealId);
+  const allTaskGroups = deal.tfm.tasks;
+
+  const newAssigneeFullName = await getAssigneeFullName(userId);
+
+  let modifiedTasks = allTaskGroups;
+
+  modifiedTasks = modifiedTasks.map((tGroup) => {
+    let group = tGroup;
+
+    group = {
+      ...group,
+      groupTasks: group.groupTasks.map((t) => {
+        let task = t;
+
+        if (teamIds.includes(task.team.id)) {
+          task = {
+            ...task,
+            assignedTo: {
+              userFullName: newAssigneeFullName,
+              userId,
+            },
+          };
+        }
+
+        return task;
+      }),
+    };
+
+    return group;
+  });
+
+  const tfmDealUpdate = {
+    tfm: {
+      tasks: modifiedTasks,
+    },
+  };
+
+  await api.updateDeal(dealId, tfmDealUpdate);
+
+  return modifiedTasks;
+};
+
+
+module.exports = assignTeamTasksToOneUser;
