+++ trade-finance-manager-api/src/v1/tasks/map-task-history-object.js
@@ -1,0 +1,20 @@
+const { getTime } = require('date-fns');
+
+const mapTaskHistoryObject = ({
+  taskId,
+  groupId,
+  statusFrom,
+  statusTo,
+  assignedUserId,
+  updatedBy,
+}) => ({
+  taskId,
+  groupId,
+  statusFrom,
+  statusTo,
+  assignedUserId,
+  updatedBy,
+  timestamp: getTime(new Date()),
+});
+
+module.exports = mapTaskHistoryObject;
